der Ball rollt wieder

+++ src/Model/Config.java
@@ -7,5 +7,5 @@ package Model;
  */
 public class Config {
 
-    public static float GRAVITY = 981f;
+    public static double GRAVITY = 981;
 }
